payment integration
Former-commit-id: e900e226a153b786fe3103c68cf1a558e01df711

+++ mobile/lib/Screens/Payment/components/body_transaction_successful.dart
@@ -4,8 +4,8 @@ import 'package:cakrawala_mobile/constants.dart';
 import 'package:flutter/material.dart';
 
 class BodyTransactionSuccessful extends StatelessWidget {
-  final double nominal;
-  final double points;
+  final int nominal;
+  final int points;
   final String namaMerchant;
   final String time;
   const BodyTransactionSuccessful({
